+++ DTCD-ExtensionRiskPrimitives/src/Abstractions/AbstractNode.js
@@ -1,0 +1,9 @@
+export default class AbstractNode {
+  instance;
+
+  constructor(yFiles) {
+    const { SimpleNode } = yFiles.default
+    this.instance = new SimpleNode();
+      this.instance.tag = {}
+  }
+}