+++ DTCD-ExtensionRiskPrimitives/src/primitives/ExportNode/ExportNode.js
@@ -1,0 +1,50 @@
+import icon from './icon.svg';
+import NodeWithDefaultLabel from '../../Abstractions/NodeWithDefaultLabel';
+
+export default class ExportNode extends NodeWithDefaultLabel {
+  static getPrimitiveInfo() {
+    return {
+      icon,
+      title: 'Экспорт данных',
+      name: 'ExportNode',
+      groups: ['Рисковые примитивы'],
+    };
+  }
+
+  constructor(yFiles) {
+    super(yFiles)
+    this.yfiles = yFiles.default;
+    this.instance.tag.defaultLabel = ``
+    this.instance.tag.defaulInitialtLabel = ``
+  }
+
+  create() {
+    const {
+      ShapeNodeStyle,
+      Rect,
+      Font,
+      MarkupLabelStyle,
+      HorizontalTextAlignment,
+      TextWrapping,
+    } = this.yfiles;
+
+
+    this.instance.layout = new Rect(0, 0, 294, 148);
+    this.instance.style = new ShapeNodeStyle({
+      shape: 'round-rectangle',
+      stroke: '4px #77dd77',
+      fill: '#fff',
+    });
+
+    const customLabelStyle = new MarkupLabelStyle({
+      font: new Font('"Segoe UI", Arial', 12),
+      horizontalTextAlignment: HorizontalTextAlignment.LEFT,
+      wrapping: TextWrapping.WORD_ELLIPSIS,
+      insets: [20, 20],
+    });
+
+    this.instance.tag = {...this.instance.tag, customLabelStyle, properties: {}, initPorts: [] };
+
+    return this.instance;
+  }
+}