+++ DTCD-ExtensionRiskPrimitives/src/primitives/FactorAnalysisNode/FactorAnalysisNode.js
@@ -1,0 +1,69 @@
+import icon from './icon.svg';
+import createNodeProperty from './../../../../DTCD-SDK/utils/createNodeProperty'
+import NodeWithDefaultLabel from '../../Abstractions/NodeWithDefaultLabel';
+
+export default class FactorAnalysisNode extends NodeWithDefaultLabel {
+  static getPrimitiveInfo() {
+    return {
+      icon,
+      title: 'Факторный анализ эффективности управления риском',
+      name: 'FactorAnalysis',
+      groups: ['Рисковые примитивы'],
+    };
+  }
+
+  constructor(yFiles) {
+    super(yFiles)
+    this.yfiles = yFiles.default;
+  }
+
+  create() {
+    const {
+      ShapeNodeStyle,
+      Rect,
+      Font,
+      MarkupLabelStyle,
+      HorizontalTextAlignment,
+      TextWrapping,
+    } = this.yfiles;
+
+
+    this.instance.layout = new Rect(0, 0, 294, 148);
+    this.instance.style = new ShapeNodeStyle({
+      shape: 'round-rectangle',
+      stroke: '4px #fde910',
+      fill: '#fff',
+    });
+
+    const customLabelStyle = new MarkupLabelStyle({
+      font: new Font('"Segoe UI", Arial', 12),
+      horizontalTextAlignment: HorizontalTextAlignment.LEFT,
+      wrapping: TextWrapping.WORD_ELLIPSIS,
+      insets: [20, 20],
+    });
+
+    const properties = {
+      type: createNodeProperty({ expression: `"Факторный анализ эффективности управления риском"`, title: 'Тип примитива' }),
+      name: createNodeProperty({title:'Наименование'}),
+      description: createNodeProperty({title:'Описание'}),
+      value: createNodeProperty({title:'Значение'}),
+      management_action: createNodeProperty({title:'Действия менеджмента'}),
+      external_influence: createNodeProperty({title:'Внешнее воздействие'}),
+    };
+
+    const initPorts = [
+      {
+        primitiveName: `inPort1`,
+        type: 'IN',
+        portPosition: { x: 0.5, y: 1 },
+        properties: {
+          status: createNodeProperty({}),
+        },
+      }
+    ];
+
+    this.instance.tag = {...this.instance.tag, customLabelStyle, properties, initPorts };
+
+    return this.instance;
+  }
+}