+++ DTCD-ExtensionRiskPrimitives/src/primitives/UncontrolledRiskNode/UncontrolledRiskNode.js
@@ -1,0 +1,76 @@
+import icon from './icon.svg';
+import createNodeProperty from './../../../../DTCD-SDK/utils/createNodeProperty'
+import NodeWithDefaultLabel from '../../Abstractions/NodeWithDefaultLabel';
+
+export default class UncontrolledRiskNode extends NodeWithDefaultLabel {
+  static getPrimitiveInfo() {
+    return {
+      icon,
+      title: 'Неподконтрольный риск-фактор (главное дерево)',
+      name: 'UncontrolledRisk',
+      groups: ['Рисковые примитивы'],
+    };
+  }
+
+  constructor(yFiles) {
+    super(yFiles)
+    this.yfiles = yFiles.default;
+  }
+
+  create() {
+    const {
+      ShapeNodeStyle,
+      Rect,
+      Font,
+      MarkupLabelStyle,
+      HorizontalTextAlignment,
+      TextWrapping,
+    } = this.yfiles;
+
+
+    this.instance.layout = new Rect(0, 0, 294, 148);
+    this.instance.style = new ShapeNodeStyle({
+      shape: 'round-rectangle',
+      stroke: '4px #CD5D67',
+      fill: '#fff',
+    });
+
+    const customLabelStyle = new MarkupLabelStyle({
+      font: new Font('"Segoe UI", Arial', 12),
+      horizontalTextAlignment: HorizontalTextAlignment.LEFT,
+      wrapping: TextWrapping.WORD_ELLIPSIS,
+      insets: [20, 20],
+    });
+
+    const properties = {
+      type: createNodeProperty({ expression: `"Неподконтрольный риск-фактор (главное дерево)"`, title: 'Тип примитива' }),
+      name: createNodeProperty({title: 'Наименование риск-фактора' }),
+      description: createNodeProperty({ title: 'Описание' }),
+      value: createNodeProperty({ title: 'Значение' }),
+      risk_clc: createNodeProperty({ title: 'Величина влияния риска' }),
+    };
+
+    const initPorts = [
+      {
+        primitiveName: `inPort1`,
+        type: 'IN',
+        portPosition: { x: 0.5, y: 1 },
+        properties: {
+          status: createNodeProperty({}),
+        },
+      },
+      {
+        primitiveName: `outPort1`,
+        type: 'OUT',
+        portPosition: { x: 0.5, y: 0 },
+        properties: {
+          status: createNodeProperty({ expression: `value`}),
+        },
+      }
+    ];
+
+    this.instance.tag = {...this.instance.tag, customLabelStyle, properties, initPorts };
+
+    return this.instance;
+  }
+}